refactor(auth/gcp): move the scrub out of client.go
Pure code motion, no line of either file changed. client.go had grown to 961
lines and more than half of them were the redaction machinery, which has
nothing to do with driving an HTTP client: reading it meant scrolling past
500 lines of string scanning to find doPost.

The two now split where the responsibility does. client.go is the credential
client — Client, Config, Request, RetrieveCredential, the outcome sum type,
doPost. redact.go is everything that turns service-controlled text into text
an error may carry, and nothing in it is specific to GCP.

Same package deliberately. There is exactly one caller of this machinery in
the repository and it is in this package, and the provider that arrives later
in the stack lands here too. A shared package would have to either impose this
package's policy on unknown callers — the withheld constant, the lowercasing,
the cap measured on the source, the 4 KiB bound — or parameterise it against a
single data point. Splitting the file makes that move a git mv when a second
caller actually shows up, without committing to a shape now.
